mgsim: fixed most conversions warnings for gcc-4.3. There are still one or two left, but they are probably wrong in gcc. gcc-4.4 will probably fix it.
git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@2158 e97e5017-a994-416e-809d-76780e9f78db

+++ src/BankedMemory.cpp
@@ -132,7 +132,7 @@ Result BankedMemory::Write(IMemoryCallback& callback, MemAddr address, void* dat
         request.data.size = size;
         request.data.tag  = tag;
         request.write     = true;
-        memcpy(request.data.data, data, size);
+        memcpy(request.data.data, data, (size_t)size);
 
         // Broadcast the snoop data
         for (set<IMemoryCallback*>::iterator p = m_caches.begin(); p != m_caches.end(); ++p)
